producer-service/requirements.txt
fastapi
uvicorn
kafka-python
pytest
psycopg2-binary
httpx
